Organize news and profile components into respective folders
