+++ client/components/news/locationBar.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
-import submitLoc from '../actions/locationBarActions';
+import submitLoc from '../../actions/locationBarActions';
 
 class LocationBar extends Component {
   constructor(props) {
