+++ client/components/news/newsItem.js
