+++ client/routes.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import { Route, IndexRoute, Link } from 'react-router';
 // import Axios from 'axios';
-import LocationBar from './components/locationBar';
+import LocationBar from './components/news/locationBar';
 import MenuContainer from './containers/sidebarContainer';
 import MapContainer from './components/events/eventsMap';
 import Events from './components/events/eventsDisplay';
-import News from './components/newsList';
+import News from './components/news/newsList';
 import PageNotFound from './components/PageNotFound';
-import Login from './components/Login';
-import Signup from './components/Signup';
-import Auth from './components/Auth';
+import Login from './components/profile/Login';
+import Signup from './components/profile/Signup';
+import Auth from './components/profile/Auth';
 import Top20 from './components/charts/top20Wrapper';
 import HeaderContainer from './containers/headerContainer';
 import HomePage from './components/Landing';
